Add markdown_area to ActionText::TagHelper and FormBuilder
