+++ app/views/pages/_form.html.erb
@@ -1,5 +1,3 @@
 <%= form_with model: [ book, page ], id: "page-editor", format: :html do |form| %>
-  <%= house_field "page[body]", page.body.content, toolbar: "house_toolbar",
-        data: { uploads_url: action_text_markdown_uploads_url(record_gid: page.to_signed_global_id.to_s, attribute_name: "body", format: "json") },
-        required: true, autofocus: true, class: "page__editor" %>
+  <%= form.markdown_area :body, toolbar: "house_toolbar", required: true, autofocus: true, class: "page__editor" %>
 <% end %>
